Moved early return tests into separate test class
