+++ debuglog-plugin/src/test/kotlin/com/bnorm/debug/log/IrPluginTest.kt
@@ -16,23 +16,10 @@
 
 package com.bnorm.debug.log
 
-import com.strobel.assembler.InputTypeLoader
-import com.strobel.assembler.metadata.ArrayTypeLoader
-import com.strobel.assembler.metadata.CompositeTypeLoader
-import com.strobel.assembler.metadata.ITypeLoader
-import com.strobel.decompiler.Decompiler
-import com.strobel.decompiler.DecompilerSettings
-import com.strobel.decompiler.PlainTextOutput
 import com.tschuchort.compiletesting.KotlinCompilation
 import com.tschuchort.compiletesting.SourceFile
-import org.jetbrains.kotlin.compiler.plugin.ComponentRegistrar
-import org.jetbrains.kotlin.utils.indexOfFirst
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
-import java.io.ByteArrayOutputStream
-import java.io.PrintStream
-import java.io.StringWriter
-import java.lang.reflect.InvocationTargetException
 
 class IrPluginTest {
 
@@ -55,39 +42,6 @@ fun greet(greeting: String = "Hello", name: String = "World"): String {
 @DebugLog
 fun doSomething() {
     Thread.sleep(15)
-}
-"""
-  )
-
-  private val earlyReturn = SourceFile.kotlin(
-    "earlyReturn.kt", """
-import com.bnorm.debug.log.DebugLog
-
-@DebugLog
-fun earlyReturn(input: String): String {
-    if (input == "EARLY_RETURN_1") {
-        return "Early return - 1"
-    }
-    if (input == "EARLY_RETURN_2") {
-        return "Early return - 2"
-    }
-    return "Normal return"
-}
-"""
-  )
-
-  private val earlyExit = SourceFile.kotlin(
-    "earlyExit.kt", """
-import com.bnorm.debug.log.DebugLog
-
-@DebugLog
-fun earlyExit(input: String) {
-    if (input == "EARLY_RETURN_1") {
-        return
-    }
-    if (input == "EARLY_RETURN_2") {
-        return
-    }
 }
 """
   )
@@ -175,174 +129,5 @@ fun earlyExit(input: String) {
     assertTrue(out.size == 2)
     assertTrue(out[0] == "Hello, World!")
     assertTrue(out[1] == "Hello, Kotlin IR!")
-  }
-
-  @Test
-  fun `IR plugin enabled - with early return`() {
-    val result = compile(sourceFile = earlyReturn, DebugLogComponentRegistrar(true))
-    assertEquals(KotlinCompilation.ExitCode.OK, result.exitCode)
-
-    assertFunction(result.javaCode("EarlyReturnKt"), "public static final String earlyReturn",
-      """
-      public static final String earlyReturn(@NotNull final String input) {
-          Intrinsics.checkNotNullParameter(input, "input");
-          System.out.println((Object)("⇢ earlyReturn(input=" + input + ')'));
-          final TimeMark markNow = TimeSource.Monotonic.INSTANCE.markNow();
-          try {
-              if (Intrinsics.areEqual(input, "EARLY_RETURN_1")) {
-                  System.out.println((Object)("⇠ earlyReturn [" + (Object)Duration.toString-impl(markNow.elapsedNow-UwyO8pc()) + "] = " + "Early return - 1"));
-                  return "Early return - 1";
-              }
-              if (Intrinsics.areEqual(input, "EARLY_RETURN_2")) {
-                  System.out.println((Object)("⇠ earlyReturn [" + (Object)Duration.toString-impl(markNow.elapsedNow-UwyO8pc()) + "] = " + "Early return - 2"));
-                  return "Early return - 2";
-              }
-              System.out.println((Object)("⇠ earlyReturn [" + (Object)Duration.toString-impl(markNow.elapsedNow-UwyO8pc()) + "] = " + "Normal return"));
-              return "Normal return";
-          }
-          catch (Throwable t) {
-              System.out.println((Object)("⇠ earlyReturn [" + (Object)Duration.toString-impl(markNow.elapsedNow-UwyO8pc()) + "] = " + t));
-              throw t;
-          }
-      }
-      """.trimIndent())
-  }
-
-  @Test
-  fun `IR plugin enabled - with early exit`() {
-    val result = compile(sourceFile = earlyExit, DebugLogComponentRegistrar(true))
-    assertEquals(KotlinCompilation.ExitCode.OK, result.exitCode)
-
-    assertFunction(result.javaCode("EarlyExitKt"), "public static final void earlyExit",
-      """
-      public static final void earlyExit(@NotNull final String input) {
-          Intrinsics.checkNotNullParameter(input, "input");
-          System.out.println((Object)("⇢ earlyExit(input=" + input + ')'));
-          final TimeMark markNow = TimeSource.Monotonic.INSTANCE.markNow();
-          try {
-              if (Intrinsics.areEqual(input, "EARLY_RETURN_1")) {
-                  System.out.println((Object)("⇠ earlyExit [" + (Object)Duration.toString-impl(markNow.elapsedNow-UwyO8pc()) + ']'));
-                  return;
-              }
-              if (Intrinsics.areEqual(input, "EARLY_RETURN_2")) {
-                  System.out.println((Object)("⇠ earlyExit [" + (Object)Duration.toString-impl(markNow.elapsedNow-UwyO8pc()) + ']'));
-                  return;
-              }
-              System.out.println((Object)("⇠ earlyExit [" + (Object)Duration.toString-impl(markNow.elapsedNow-UwyO8pc()) + ']'));
-          }
-          catch (Throwable t) {
-              System.out.println((Object)("⇠ earlyExit [" + (Object)Duration.toString-impl(markNow.elapsedNow-UwyO8pc()) + "] = " + t));
-              throw t;
-          }
-      }
-      """.trimIndent())
-  }
-
-  private fun assertFunction(javaCode: String, functionStatement: String, expectedFunction: String) {
-    assertEquals(expectedFunction, fetchMethodByPrefix(javaCode, functionStatement))
-  }
-
-  private fun fetchMethodByPrefix(classText: String, methodSignaturePrefix: String): String {
-    val classLines = classText.split("\n")
-    val methodSignaturePredicate: (String) -> Boolean = { line -> line.contains(methodSignaturePrefix) }
-    val methodFirstLineIndex = classLines.indexOfFirst(methodSignaturePredicate)
-
-    check(methodFirstLineIndex != -1) {
-      "Method with prefix '$methodSignaturePrefix' not found within class:\n$classText"
-    }
-
-    val multiplePrefixMatches = classLines
-      .indexOfFirst(methodFirstLineIndex + 1, methodSignaturePredicate)
-      .let { index -> index != -1 }
-
-    check(!multiplePrefixMatches) {
-      "Multiple methods with prefix '$methodSignaturePrefix' found within class:\n$classText"
-    }
-
-    val indentationSize = classLines[methodFirstLineIndex].takeWhile { it == ' ' }.length
-
-    var curleyBraceCount = 1
-    var currentLineIndex: Int = methodFirstLineIndex + 1
-
-    while (curleyBraceCount != 0 && currentLineIndex < classLines.lastIndex) {
-      if (classLines[currentLineIndex].contains("{")) {
-        curleyBraceCount++
-      }
-      if (classLines[currentLineIndex].contains("}")) {
-        curleyBraceCount--
-      }
-      currentLineIndex++
-    }
-
-    return classLines
-      .subList(methodFirstLineIndex, currentLineIndex)
-      .joinToString("\n") { it.substring(indentationSize) }
-  }
-}
-
-fun compile(
-  sourceFiles: List<SourceFile>,
-  plugin: ComponentRegistrar,
-): KotlinCompilation.Result {
-  return KotlinCompilation().apply {
-    sources = sourceFiles
-    useIR = true
-    compilerPlugins = listOf(plugin)
-    inheritClassPath = true
-    verbose = false
-  }.compile()
-}
-
-fun compile(
-  sourceFile: SourceFile,
-  plugin: ComponentRegistrar,
-): KotlinCompilation.Result {
-  return compile(listOf(sourceFile), plugin)
-}
-
-fun invokeMain(result: KotlinCompilation.Result, className: String): String {
-  val oldOut = System.out
-  try {
-    val buffer = ByteArrayOutputStream()
-    System.setOut(PrintStream(buffer, false, "UTF-8"))
-
-    try {
-      val kClazz = result.classLoader.loadClass(className)
-      val main = kClazz.declaredMethods.single { it.name == "main" && it.parameterCount == 0 }
-      main.invoke(null)
-    } catch (e: InvocationTargetException) {
-      throw e.targetException
-    }
-
-    return buffer.toString("UTF-8")
-  } finally {
-    System.setOut(oldOut)
-  }
-}
-
-fun KotlinCompilation.Result.javaCode(className: String): String {
-  val decompilerSettings = DecompilerSettings.javaDefaults().apply {
-    typeLoader = CompositeTypeLoader(*(mutableListOf<ITypeLoader>()
-      .apply {
-        // Ensure every class is available.
-        generatedFiles.forEach {
-          add(ArrayTypeLoader(it.readBytes()))
-        }
-
-        // Loads any standard classes already on the classpath.
-        add(InputTypeLoader())
-      }
-      .toTypedArray()))
-
-    isUnicodeOutputEnabled = true
-  }
-
-  return StringWriter().let { writer ->
-    Decompiler.decompile(
-      className,
-      PlainTextOutput(writer).apply { isUnicodeOutputEnabled = true },
-      decompilerSettings
-    )
-    writer.toString().trimEnd().trimIndent()
   }
 }
